+++ sdk/azidentity/CHANGELOG.md
@@ -10,6 +10,7 @@
 * `ManagedIdentityCredential.GetToken` no longer mutates its `opts.Scopes`
 
 ### Other Changes
+* Bumps version of `azcore` to `v0.18.1`
 
 
 ## v0.9.2 (2021-07-23)